+++ src/main/kotlin/com/hedvig/underwriter/service/quotesSignDataStrategies/RedirectSignStrategy.kt
@@ -5,10 +5,9 @@ import com.hedvig.underwriter.model.SignSessionRepository
 import com.hedvig.underwriter.service.model.StartSignErrors
 import com.hedvig.underwriter.service.model.StartSignResponse
 import com.hedvig.underwriter.serviceIntegration.memberService.MemberService
-import com.hedvig.underwriter.serviceIntegration.memberService.NationalIdentification
-import com.hedvig.underwriter.serviceIntegration.memberService.RedirectCountry
-import com.hedvig.underwriter.serviceIntegration.memberService.UnderwriterStartSignSessionRequest
-import com.hedvig.underwriter.serviceIntegration.memberService.UnderwriterStartSignSessionResponse
+import com.hedvig.underwriter.serviceIntegration.memberService.dtos.NationalIdentification
+import com.hedvig.underwriter.serviceIntegration.memberService.dtos.RedirectCountry
+import com.hedvig.underwriter.serviceIntegration.memberService.dtos.UnderwriterStartSignSessionResponse
 import com.hedvig.underwriter.serviceIntegration.memberService.dtos.Nationality
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
