+++ src/main/kotlin/com/hedvig/underwriter/serviceIntegration/memberService/MemberServiceClient.kt
@@ -1,16 +1,14 @@
 package com.hedvig.underwriter.serviceIntegration.memberService
 
-import com.fasterxml.jackson.annotation.JsonSubTypes
-import com.fasterxml.jackson.annotation.JsonTypeInfo
 import com.hedvig.underwriter.serviceIntegration.memberService.dtos.FinalizeOnBoardingRequest
 import com.hedvig.underwriter.serviceIntegration.memberService.dtos.HelloHedvigResponseDto
 import com.hedvig.underwriter.serviceIntegration.memberService.dtos.InternalMember
 import com.hedvig.underwriter.serviceIntegration.memberService.dtos.IsMemberAlreadySignedResponse
 import com.hedvig.underwriter.serviceIntegration.memberService.dtos.IsSsnAlreadySignedMemberResponse
-import com.hedvig.underwriter.serviceIntegration.memberService.dtos.Nationality
-import com.hedvig.underwriter.serviceIntegration.memberService.dtos.RedirectAuthenticationResponseError
 import com.hedvig.underwriter.serviceIntegration.memberService.dtos.PersonStatusDto
 import com.hedvig.underwriter.serviceIntegration.memberService.dtos.UnderwriterQuoteSignResponse
+import com.hedvig.underwriter.serviceIntegration.memberService.dtos.UnderwriterStartSignSessionRequest
+import com.hedvig.underwriter.serviceIntegration.memberService.dtos.UnderwriterStartSignSessionResponse
 import com.hedvig.underwriter.serviceIntegration.memberService.dtos.UpdateSsnRequest
 import com.hedvig.underwriter.web.dtos.UnderwriterQuoteSignRequest
 import feign.Headers
@@ -22,7 +20,6 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestHeader
 import org.springframework.web.bind.annotation.RequestMapping
-import java.util.UUID
 
 @Headers("Accept: application/json;charset=utf-8")
 @FeignClient(
@@ -71,73 +68,4 @@ interface MemberServiceClient {
     fun getMember(
         @PathVariable("memberId") memberId: Long
     ): ResponseEntity<InternalMember>
-}
-
-
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
-@JsonSubTypes(
-    JsonSubTypes.Type(value = UnderwriterStartSignSessionRequest.SwedishBankId::class, name = "SwedishBankId"),
-    JsonSubTypes.Type(value = UnderwriterStartSignSessionRequest.BankIdRedirect::class, name = "BankIdRedirect"),
-    JsonSubTypes.Type(value = UnderwriterStartSignSessionRequest.SimpleSign::class, name = "SimpleSign")
-)
-sealed class UnderwriterStartSignSessionRequest {
-
-    abstract val underwriterSessionReference: UUID
-
-    data class SwedishBankId(
-        override val underwriterSessionReference: UUID,
-        val nationalIdentification: NationalIdentification,
-        val ipAddress: String,
-        val isSwitching: Boolean
-    ) : UnderwriterStartSignSessionRequest()
-
-    data class BankIdRedirect(
-        override val underwriterSessionReference: UUID,
-        val nationalIdentification: NationalIdentification,
-        val successUrl: String,
-        val failUrl: String,
-        val country: RedirectCountry
-    ) : UnderwriterStartSignSessionRequest()
-
-    data class SimpleSign(
-        override val underwriterSessionReference: UUID,
-        val nationalIdentification: NationalIdentification
-    ) : UnderwriterStartSignSessionRequest()
-}
-
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
-@JsonSubTypes(
-    JsonSubTypes.Type(value = UnderwriterStartSignSessionResponse.SwedishBankId::class, name = "SwedishBankId"),
-    JsonSubTypes.Type(value = UnderwriterStartSignSessionResponse.BankIdRedirect::class, name = "BankIdRedirect"),
-    JsonSubTypes.Type(value = UnderwriterStartSignSessionResponse.SimpleSign::class, name = "SimpleSign")
-)
-sealed class UnderwriterStartSignSessionResponse {
-
-    abstract val internalErrorMessage: String?
-
-    data class SwedishBankId(
-        val autoStartToken: String?,
-        override val internalErrorMessage: String? = null
-    ) : UnderwriterStartSignSessionResponse()
-
-    data class BankIdRedirect(
-        val redirectUrl: String?,
-        override val internalErrorMessage: String? = null,
-        val errorMessages: List<RedirectAuthenticationResponseError>? = null
-    ) : UnderwriterStartSignSessionResponse()
-
-    data class SimpleSign(
-        val successfullyStarted: Boolean,
-        override val internalErrorMessage: String? = null
-    ) : UnderwriterStartSignSessionResponse()
-}
-
-data class NationalIdentification(
-    val identification: String,
-    val nationality: Nationality
-)
-
-enum class RedirectCountry {
-    NORWAY,
-    DENMARK
 }
