+++ src/main/kotlin/com/hedvig/underwriter/serviceIntegration/memberService/dtos/NationalIdentification.kt
@@ -1,0 +1,6 @@
+package com.hedvig.underwriter.serviceIntegration.memberService.dtos
+
+data class NationalIdentification(
+    val identification: String,
+    val nationality: Nationality
+)
