+++ src/main/kotlin/com/hedvig/underwriter/serviceIntegration/memberService/dtos/RedirectCountry.kt
@@ -1,0 +1,6 @@
+package com.hedvig.underwriter.serviceIntegration.memberService.dtos
+
+enum class RedirectCountry {
+    NORWAY,
+    DENMARK
+}
